requirements.txt
jira
pconf>=1.7.0
python-dateutil
